+++ src/gui/web-components/form/MetarImportWebComponent.ts
@@ -1,3 +1,4 @@
+import type { AppState } from "../../renderer/AppState.js";
 import { dispatchNotificationEvent, type NotificationEventPayload } from "../../renderer/notificationEventHandler.js";
 import { sendToMain } from "../../renderer/sendToMain.js";
 import { AbstractStateSubscriberWebComponent } from "../util/AbstractStateSubscriberWebComponent.js";
@@ -10,6 +11,7 @@ export class MetarImportWebComponent extends AbstractStateSubscriberWebComponent
     private isInitialized = false;
 
     private elements!: {
+        metarButton: HTMLButtonElement;
         metarOrigin: HTMLButtonElement;
         metarDestination: HTMLButtonElement;
         dialog: HTMLDialogElement;
@@ -38,6 +40,7 @@ export class MetarImportWebComponent extends AbstractStateSubscriberWebComponent
 </dialog>
         `;
         this.elements = {
+            metarButton: this.querySelector("button") as HTMLButtonElement,
             metarOrigin: this.querySelector("#metar-origin") as HTMLButtonElement,
             metarDestination: this.querySelector("#metar-destination") as HTMLButtonElement,
             dialog: this.querySelector("dialog") as HTMLDialogElement,
@@ -50,6 +53,7 @@ export class MetarImportWebComponent extends AbstractStateSubscriberWebComponent
             this.isInitialized = true;
         }
         this.subscribeToStateUpdates((state) => {
+            this.elements.metarButton.disabled = this.isButtonDisabled(state);
             this.elements.metarOrigin.innerHTML = state.route.departureAirport || "Origin";
             this.elements.metarOrigin.dataset.icao = state.route.departureAirportCode || "";
             this.elements.metarDestination.disabled = !state.route.destinationAirportCode;
@@ -73,6 +77,16 @@ export class MetarImportWebComponent extends AbstractStateSubscriberWebComponent
         super.disconnectedCallback();
         this.elements.metarOrigin.removeEventListener("click", this.handleClickOrigin);
         this.elements.metarDestination.removeEventListener("click", this.handleClickDestination);
+    }
+
+    private isButtonDisabled(state: AppState): boolean {
+        const date = new Date(state.dateTime.utc.date + "T" + state.dateTime.utc.time + "Z");
+        const fourWeeksAgo = new Date();
+        fourWeeksAgo.setDate(fourWeeksAgo.getDate() - 28);
+        if (date < fourWeeksAgo || date > new Date()) {
+            return true;
+        }
+        return !state.route.departureAirportCode && !state.route.destinationAirportCode;
     }
 
     private handleClickOrigin = () => {